feat(UM-91): refactor Root and AdminRoot controllers and assemblers
