+++ apps/api/src/main/java/dev/kurama/api/core/hateoas/root/model/RootResource.java
@@ -1,15 +1,8 @@
-package dev.kurama.api.core.hateoas.model;
+package dev.kurama.api.core.hateoas.root.model;
 
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import org.springframework.hateoas.RepresentationModel;
 
-@EqualsAndHashCode(callSuper = true)
-@Data
-@Builder
 @NoArgsConstructor
 public class RootResource extends RepresentationModel<RootResource> {
-
 }
