+++ apps/api/src/main/java/dev/kurama/api/core/hateoas/root/processor/RootAssembler.java
@@ -1,0 +1,9 @@
+package dev.kurama.api.core.hateoas.root.processor;
+
+import lombok.NonNull;
+import org.springframework.hateoas.RepresentationModel;
+
+public interface RootAssembler<T extends RepresentationModel<T>> {
+
+  @NonNull RepresentationModel<T> assemble();
+}
